Add messages for motion primitives
These messages should reflect the motion primitives supported by most industrial
robots.They are meant to being extended in the future in order to support
a larger variety of motion implementations.

The motivation is to support executing robot motions as they are
traditionally programmed: A series of primitives such as `LIN`,
`PTP`, `CIRC`.